+++ CHANGES.md
@@ -1,5 +1,15 @@
 # IndexedDBShim changes
 
+## ?
+
+- Fix: For interface tests, ensure properties are getter based
+- Refactoring: Factor out common interface code (making
+    `IDBRequest` and `IDBOpenRequest` listeners `enumerable` too)
+- Testing (W3C): Update web-platform-tests (including repo location)
+- Testing (W3C): Use new idlharness.any.html over interfaces.any.html
+- Travis: Update Node version (having issues with Yarn)
+- Docs: Remove Bower badge
+
 ## 3.9.0
 
 - Fix: Upon converting value to key range, avoid creating new